requirements.txt
bottle
gunicorn
pyyaml